Restore bugfixes tests (#637)
