+++ test/ColunaTests.jl
@@ -48,7 +48,7 @@ module ColunaTests
     # ########################################################################################
     # # Bugfixes tests
     # ########################################################################################
-    # include("issues_tests.jl")
+    include("bugfixes.jl")
 
     ########################################################################################
     # Other tests
